wip: Move stash db queries into utils
